requirements.txt
pandas
numpy
scikit-learn
mlflow
fastapi
uvicorn
pydantic
loguru
python-box
matplotlib
seaborn
pyyaml
ensure
joblib
dill
pytest


